backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn
gunicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
requests
aiohttp

# Auth
PyJWT
python-jose[cryptography]
passlib[bcrypt]

# AI & News
google-generativeai
feedparser
edge-tts

# Video
ffmpeg-python

# AI Lip-Sync (SadTalker/Wav2Lip)
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2
numba
resampy==0.3.1
pydub==0.25.1
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gfpgan
av
safetensors
opencv-python-headless
